Rework with stable sorting-network
With hindsight it's kind of obvious that swapping non-adjacent wires
would lead to an unstable sort. But I had a test for that and it
didn't trigger, even though it has done so in the past.
This uses a transposition sort8 together with the novel merge into
approach. For sizes <= 20 a simpler version was built with better
cold code performance and reverse pattern sorting was fixed.