Avoid picking pivot that is the smallest element
Previously it was possible and likely for inputs of low cardinality to pick a
pivot that had no smaller element in the slice. Such pivot is uniquely bad as
the following partition will attempt to split it into elements smaller and equal
or greater. The smaller partition will have a size of zero. Essentially
completely wasting the partitioning operation. By adding code that detects such
cases and choosing a pivot that is more than the previously chosen value, we
avoid such degenerate cases. This improves the performance for inputs with low
cardinality, peaking at 2x for binary distributions.