Avoid was_partitioned search for relatively small slices
There is a small area of slice length that is very common, and the initial
branchy searching has ~5% perf impact on random inputs. By not doing this search
for a small extended area after the area covered by small sort, we improve many
benchmarks, while minimizing the chances of not hitting was_partitioned +
partial_insertion_sort exit. Average worst case one additional portioning has to
be done, if that very specific size hole is hit with a sub-slice.