Refine optimization selection in ipn_stable
Allow the use of parity_merge regardless of the type size.
It's much more efficient than the regular merge function,
even if the sort networks can't be used.

Allow small sort up to 32, for all types that are cheap to move
regardless if they have direct interior mutability or not.