Re-tune ipn_stable small sort for cold code
Avoid perf regressions compared to rust_std_stable by carfully putting
an inlined insertion sort in the hot path. Re-tune heuristics for when
stable sorting-networks are worth it. Use a different pattern analysis
algorithm, that will respect a pre-sorted streak in small-sort if it is
the largest part of the input.

Also improve edge run creation.