Add random_cell_i32 test
In addition to other tests that check for observation safety of types with
interior mutability. This new test, tests that such types produce a correct
sorted output.

starting # with '#' will be ignored, and an empty message aborts the commit. # #
On branch main # Your branch is up to date with 'origin/main'. # # Changes to be
committed: #	modified:   sort_test_tools/src/tests.rs #