Switch ipnsort to new partition implementation
Many months of frustrating research finally pay of with
lomuto_branchless_cyclic_opt. It is very fast on Zen3 and Firestorm, for u64,
f128 and String. Has tiny binary-size and compile-time impact. Is easy to reason
about and works well for many types. Expects little of the prefetcher. Paired
with a fallback minimal-copy partition for large types.

Die 🌰 wurde geknackt!