Restore branchless code-gen of rust_std_stable
Until LLVM 16 the code in the `slice::sort` `merge` function generated
branchless code. With the upgrade to LLVM 16 in rustc 1.70 this property was
broken. See https://github.com/rust-lang/rust/issues/111559 for more info. It's
annoying for comparison reasons to have the same code but with significantly
different code-gen with newer versions. The updated merge function generates
pretty much the same code as the previous one did with older rustc versions.